Tests/modernize (#8445)
* working

* cleanfirst

* fixing bug in CMakeDeps
