+++ conan/tools/cmake/cmakedeps.py
@@ -358,7 +358,7 @@ endforeach()
     components_target_build_type_tpl = Template(textwrap.dedent("""\
         ########## MACROS ###########################################################################
         #############################################################################################
-        include(${{CMAKE_CURRENT_LIST_DIR}}/cmakedeps_macros.cmake)
+        include(${CMAKE_CURRENT_LIST_DIR}/cmakedeps_macros.cmake)
 
         ########### VARIABLES #######################################################################
         #############################################################################################
