+++ conans/test/functional/generators/cmake_find_package_multi_test.py
@@ -4,7 +4,6 @@ import textwrap
 import unittest
 
 import pytest
-from parameterized import parameterized
 
 from conans.model.ref import ConanFileReference, PackageReference
 from conans.test.assets.sources import gen_function_cpp
@@ -355,93 +354,6 @@ class CMakeFindPathMultiGeneratorTest(unittest.TestCase):
                       "$<$<STREQUAL:$<TARGET_PROPERTY:TYPE>,MODULE_LIBRARY>:>;"
                       "$<$<STREQUAL:$<TARGET_PROPERTY:TYPE>,EXECUTABLE>:>>",
                       client.out)
-
-    def test_no_version_file(self):
-        client = TestClient()
-        client.run("new hello/1.1 -s")
-        client.run("create .")
-
-        cmakelists = textwrap.dedent("""
-            cmake_minimum_required(VERSION 3.1)
-            project(consumer)
-            find_package(hello 1.0 REQUIRED)
-            message(STATUS "hello found: ${{hello_FOUND}}")
-            """)
-
-        conanfile = textwrap.dedent("""
-            from conans import ConanFile, CMake
-
-            class Conan(ConanFile):
-                settings = "build_type"
-                requires = "hello/1.1"
-                generators = "cmake_find_package_multi"
-
-                def build(self):
-                    cmake = CMake(self)
-                    cmake.configure()
-            """)
-
-        client.save({"conanfile.py": conanfile, "CMakeLists.txt": cmakelists})
-        client.run("install .")
-        os.unlink(os.path.join(client.current_folder, "hello-config-version.cmake"))
-        exit_code = client.run("build .", assert_error=True)
-        self.assertNotEqual(0, exit_code)
-
-    @parameterized.expand([
-        ("hello", "1.0", "", False, True),
-        ("Hello", "1.0", "", False, True),
-        ("HELLO", "1.0", "", False, True),
-        ("hello", "1.1", "", False, True),
-        ("hello", "1.2", "", False, False),
-        ("hello", "1.0", "EXACT", False, False),
-        ("hello", "1.1", "EXACT", False, True),
-        ("hello", "1.2", "EXACT", False, False),
-        ("hello", "0.1", "", False, False),
-        ("hello", "2.0", "", False, False),
-        ("hello", "1.0", "REQUIRED", False, True),
-        ("hello", "1.1", "REQUIRED", False, True),
-        ("hello", "1.2", "REQUIRED", True, False),
-        ("hello", "1.0", "EXACT REQUIRED", True, False),
-        ("hello", "1.1", "EXACT REQUIRED", False, True),
-        ("hello", "1.2", "EXACT REQUIRED", True, False),
-        ("hello", "0.1", "REQUIRED", True, False),
-        ("hello", "2.0", "REQUIRED", True, False)
-    ])
-    def test_version(self, name, version, params, cmake_fails, package_found):
-        client = TestClient()
-        client.run("new hello/1.1 -s")
-        client.run("create .")
-
-        cmakelists = textwrap.dedent("""
-            cmake_minimum_required(VERSION 3.1)
-            project(consumer)
-            find_package({name} {version} {params})
-            message(STATUS "hello found: ${{{name}_FOUND}}")
-            """).format(name=name, version=version, params=params)
-
-        conanfile = textwrap.dedent("""
-            from conans import ConanFile, CMake
-
-
-            class Conan(ConanFile):
-                settings = "build_type"
-                requires = "hello/1.1"
-                generators = "cmake_find_package_multi"
-
-                def build(self):
-                    cmake = CMake(self)
-                    cmake.configure()
-            """)
-
-        client.save({"conanfile.py": conanfile, "CMakeLists.txt": cmakelists})
-        client.run("install .")
-        exit_code = client.run("build .", assert_error=cmake_fails)
-        if cmake_fails:
-            self.assertNotEqual(exit_code, 0)
-        elif package_found:
-            self.assertIn("hello found: 1", client.out)
-        else:
-            self.assertIn("hello found: 0", client.out)
 
     def test_cpp_info_config(self):
         conanfile = textwrap.dedent("""
